further refactor highlighting
